refactor: port useNormalizedInputProps to typescript
